fix: dead code 제거 및 holdings 타입 불일치 수정
- logic.py: 사용되지 않는 val_c_total 변수 및 관련 코드 제거
- models.py: Portfolio.holdings 타입을 Dict[str, float] → Dict[str, int]로 수정 (Order.quantity: int와 일치)
- broker.py: MockBroker.holdings 타입도 동일하게 수정

+++ src/core/models.py
@@ -28,7 +28,7 @@ class MarketData:
 class Portfolio:
     """현재 계좌 상태"""
     total_cash: float
-    holdings: Dict[str, float]       # {ticker: quantity}
+    holdings: Dict[str, int]          # {ticker: quantity}
     current_prices: Dict[str, float] # {ticker: price}
 
     @property
